pyproject.toml
[project]
name = "fhir-dicom-mcp"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.3.0",
    "pynetdicom>=2.1.1",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "mysql-connector-python>=9.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "pydicom>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=6.0.0",
    "reportlab"
]
[project.scripts]
dicom-mcp = "dicom_mcp.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
